+++ gengora/server/Constants.cs
@@ -40,6 +40,8 @@ internal static class Constants
         public const string GENERATOR_HELLO = "generator/hello";
         public const string GENERATOR_PUBLISH_DIAGNOSTICS = "generator/publishDiagnostics";
         public const string GENERATOR_PUBLISH_DIAGNOSTICS_ALT = "generator.publishDiagnostics";
+        public const string OBSERVATION_MODE_CHANGED = "generator/observationModeChanged";
+        public const string GENERATOR_PROJECT_DISCOVERED = "generator/projectDiscovered";
     }
 
     /// <summary>